Made controller LeftFeedback, fix issue
