+++ app/views/left_feedback/new.html.slim
@@ -1,0 +1,26 @@
+h1.pricing-h Contact us
+p.pricing-p You can call us or find our contacts
+.container.mb-5
+  .row
+    .col-md-6
+      = simple_form_for @feedback, url: left_feedback_index_path, method: 'post' do |f|
+        = f.input :name
+        = f.input :email
+        = f.input :message, as: :text, input_html:{ rows: '14' }
+        = f.button :submit, class: 'form-control btn btn-primary'
+    .col-md-6
+      .col.text-center.border.rounded.p-2.mr-md-2.m-1
+        i.bx.bx-phone-call.primary.font-large-1.text-muted.p-50
+        h5
+          a href="tel:+81025482568" + (810) 2548 2568
+        p.text-muted.font-medium-1 call to us
+      .col.text-center.border.rounded.p-2.mr-md-2.m-1
+        i.bx.bx-mail-send.primary.font-large-1.p-50
+        h5
+          a href="mail:remote-flow@mail.com" remote-flow@mail.com
+        p.text-muted.font-medium-1 write to us
+      .col.text-center.border.rounded.p-2.mr-md-2.m-1
+        i.bx.bxs-map.primary.font-large-1.text-muted.p-50
+        h5
+          a href="https://goo.gl/maps/GBbjjV8jCERVwhaE6" Akademika Sakharova Street, 23z, Ivano-Frankivsk, Ivano-Frankivsk Oblast
+        p.text-muted.font-medium-1 visit us
