+++ app/views/shared/public/_header.html.slim
@@ -5,7 +5,7 @@
   nav.my-2.my-md-0.mr-md-3
     = link_to 'About Us', about_us_path, class: 'p-2 text-dark'
     a.p-2.text-dark href="#"  How To Use
-    = link_to 'Contact Us', contact_us_path, class: 'p-2 text-dark'
+    = link_to 'Contact Us', new_left_feedback_path, class: 'p-2 text-dark'
     a.p-2.text-dark href="#"  Support
     = link_to 'Pricing', pricing_path, class: 'p-2 text-dark'
     - if user_signed_in?
