+++ config/routes.rb
@@ -12,10 +12,11 @@ Rails.application.routes.draw do
 
   get '/pricing', to: 'public#pricing'
   get '/about_us', to: 'public#about_us'
-  get '/contact_us', to: 'public#contact_us'
-  post '/contact_us', to: 'public#feedback'
+  # get '/contact_us', to: 'public#contact_us'
+  # post '/contact_us', to: 'public#feedback'
 
   resources :accept_invites, only: %i[new create]
+  resources :left_feedback, only: %i[new create]
 
   namespace :account do
     get '/', to: 'dashboard#index'
